Wireframe (finished members, started sales)
Added some wireframes.

The basic outline for member is done (though could likely be improved upon). I've also started doing sales using the same structure.

I originally started using a web app for the wireframe, but didn't realise it was limited to 3 pages, so I've gone back to good old paint now.

I've included some folders to make hopefully make the wireframes a bit neater and to give a sense as to the logic, alongside a diagram that shows the user flow.